Add library to mdbook crate
