+++ .github/workflows/main.yml
@@ -122,7 +122,8 @@ jobs:
           curl -Lf https://github.com/obi1kenobi/cargo-semver-checks/releases/download/v0.48.0/cargo-semver-checks-x86_64-unknown-linux-gnu.tar.gz \
             | tar -xz --directory=./installed-bins
           echo `pwd`/installed-bins >> $GITHUB_PATH
-      - run: cargo semver-checks --workspace
+      # TODO: drop `--exclude mdbook` after the first published `mdbook` lib
+      - run: cargo semver-checks --workspace --exclude mdbook
 
   # The success job is here to consolidate the total success/failure state of
   # all other jobs. This job is then included in the GitHub branch protection
